fix: 更新测试用例
